Bernoulli {BAS}	R Documentation

Independent Bernoulli Prior Distribution for Models

Description

Creates an object representing the prior distribution on models for BAS.

Usage


Bernoulli(probs = 0.5)

Arguments

probs
a scalar or vector of prior inclusion probabilities. If a scalar, the values is replicated for all variables ans a 1 is added for the intercept. BAS checks to see if the length is equal to the dimension of the parameter vector for the full model and adds a 1 to include the intercept.

Details

The independent Bernoulli prior distribution is a commonly used prior in BMA, with the Uniform distribution a special case with probs=.5. If all indicator variables have a independent Bernoulli distributions with common probability probs, the distribution on model size binomial(p, probs) distribution.

Value

returns an object of class "prior", with the family and hyperparameters.

Author(s)

Merlise Clyde

See Also

bas.lm, beta.binomial,uniform

Other priors modelpriors: Bernoulli.heredity, beta.binomial, tr.beta.binomial, tr.poisson, tr.power.prior, uniform

Examples


Bernoulli(.9)

[Package BAS version 1.5.4 Index]