Bernoulli.heredity {BAS}	R Documentation

Independent Bernoulli prior on models that with constraints for model hierarchy induced by interactions

Description

Independent Bernoulli prior on models that with constraints for model hierarchy induced by interactions

Usage


Bernoulli.heredity(pi = 0.5, parents)

Arguments

pi
Bernoulli probability that term is included

parents
matrix of terms and parents with indicators of which terms are parents for each term

Note

Not implemented yet for use with bas.lm or bas.glm

See Also

Other priors modelpriors: Bernoulli, beta.binomial, tr.beta.binomial, tr.poisson, tr.power.prior, uniform

[Package BAS version 1.5.4 Index]