CCH {BAS}	R Documentation

Generalized g-Prior Distribution for Coefficients in BMA Models

Description

Creates an object representing the CCH mixture of g-priors on coefficients for BAS .

Usage


CCH(alpha, beta, s = 0)

Arguments

alpha
a scalar > 0, recommended alpha=.5 (betaprime) or 1 for CCH. The hyper.g(alpha) is equivalent to CCH(alpha -2, 2, 0). Liang et al recommended values in the range 2 < alpha_h <= 4

beta
a scalar > 0. The value is not updated by the data; beta should be a function of n for consistency under the null model. The hyper-g corresponds to b = 2

s
a scalar, recommended s=0

Details

Creates a structure used for bas.glm.

Value

returns an object of class "prior", with the family and hyperparameters.

Author(s)

Merlise A Clyde

See Also

IC.prior, bic.prior, bas.glm

Other beta priors: EB.local, IC.prior, Jeffreys, TG, beta.prime, g.prior, hyper.g.n, hyper.g, intrinsic, robust, tCCH, testBF.prior

Examples


CCH(alpha = .5, beta = 100, s = 0)

[Package BAS version 1.5.4 Index]