IC.prior {BAS}	R Documentation

Information Criterion Families of Prior Distribution for Coefficients in BMA Models

Description

Creates an object representing the prior distribution on coefficients for BAS.

Usage


IC.prior(penalty)

Arguments

penalty
a scalar used in the penalized loglikelihood of the form penalty*dimension

Details

The log marginal likelihood is approximated as -2*(deviance + penalty*dimension). Allows alternatives to AIC (penalty = 2) and BIC (penalty = log(n)). For BIC, the argument may be missing, in which case the sample size is determined from the call to 'bas.glm' and used to determine the penalty.

Value

returns an object of class "prior", with the family and hyerparameters.

Author(s)

Merlise Clyde

See Also

g.prior

Other beta priors: CCH, EB.local, Jeffreys, TG, beta.prime, g.prior, hyper.g.n, hyper.g, intrinsic, robust, tCCH, testBF.prior

Examples


IC.prior(2)
aic.prior()
bic.prior(100)

[Package BAS version 1.5.3 Index]