Jeffreys Prior Distribution for $g$ for Mixtures of g-Priors for Coefficients in BMA Models

Creates an object representing the Jeffrey's Prior on g mixture of g-priors on coefficients for BAS. This is equivalent to a limiting version of the CCH(a, 2, 0) with a = 0 or they hyper-g(a = 2) and is an improper prior. As $g$ does not appear in the Null Model, Bayes Factors and model probabilities are not well-defined because of arbitrary normalizing constants, and for this reason the null model is excluded and the same constants are used across other models.

Jeffreys()

Value

returns an object of class "prior", with the family and hyerparameters.

Details

Creates a structure used for bas.glm.

See also

Examples

Jeffreys()

#> $family
#> [1] "Jeffreys"
#> 
#> $class
#> [1] "TCCH"
#> 
#> $hyper.parameters
#> $hyper.parameters$alpha
#> [1] 0
#> 
#> $hyper.parameters$beta
#> [1] 2
#> 
#> $hyper.parameters$s
#> [1] 0
#> 
#> 
#> attr(,"class")
#> [1] "prior"