Horseshoe Crab Data

Data on horseshoe crabs (Limulus polyphemus). Responseis number of males surrounding a breeding female, color (factor), condition (factor), weight (quantitative), and width (quantitative) of the female.

Format

A data frame with 173 observations on 6 variables. Individuals (rows of the data frame) are female horseshoe crabs. Variables other than satell refer to these females. The variables are

color

color. The colors given in Agresti are “light medium”, “medium”, “dark medium”, and “dark”. Here they are abbreviated to light, medium, dark, and darker, respectively.

spine

spine condition. The conditions given in Agresti are “both good”, “one worn or broken”, and “both worn or broken”. Here they are abbreviated to good, middle, bad, respectively.

width

carapace width in centimeters

satell

number of satellites, which males clustering around the female in addition to the male with which she is breeding.

weight

weight in grams.

y

shorthand for as.numeric(satell > 0).

Source

Agresti, A. (2013) Categorical Data Analysis, Wiley, Hoboken, NJ., Section 4.3.2, http://www.stat.ufl.edu/~aa/cda/data.html

Brockmann, H. J. (1996) Satellite Male Groups in Horseshoe Crabs, Limulus polyphemus, Ethology, 102, 1--21.

Details

Quoting from the abstract of Brockmann (1996). “Horseshoe crabs arrive on the beach in pairs and spawn ... during ... high tides. Unattached males also come to the beach, crowd around the nesting couples and compete with attached males for fertilizations. Satellite males form large groups around some couples while ignoring others, resulting in a nonrandom distribution that cannot be explained by local environmental conditions or habitat selection.”