Extract the Median Probability Model

Extracts the Median Probability Model from a bas object

extract_MPM(object)

Arguments

object
An object of class "bas" or "basglm"

Value

a new object with of class "bas" or "basglm" with the Median Probability Model

Details

The Median Probability Model is the model where variables are included if the marginal posterior probabilty of the coefficient being zero is greater than 0.5. As this model may not have been sampled (and even if it has) it is oftern faster to refit the model using bas, rather than search the list of models to see where it was included.

See also

Examples

data(Hald, package="BAS")
hald_bic =  bas.lm(Y ~ ., data=Hald, alpha=13, prior="BIC")
extract_MPM(hald_bic)
#> 
#> Call:
#> bas.lm(formula = Y ~ ., data = Hald, prior = "BIC", alpha = 13, 
#>     n.models = 1, bestmodel = c(1, 1, 1, 1, 1))
#> 
#> 
#>  Marginal Posterior Inclusion Probabilities: 
#> Intercept         X1         X2         X3         X4  
#>    1.0000     0.9870     0.8252     0.6078     0.6404  

data(Pima.tr, package="MASS")
Pima_bas = bas.glm(type ~ ., data=Pima.tr, n.models= 2^7, method="BAS",
           betaprior=CCH(a=1, b=nrow(Pima.tr)/2, s=0), family=binomial(),
           modelprior=uniform())
extract_MPM(Pima_bas)
#> 
#> Call:
#> bas.glm(formula = type ~ ., family = binomial(), data = Pima.tr, 
#>     n.models = 1, betaprior = CCH(a = 1, b = nrow(Pima.tr)/2, 
#>         s = 0), modelprior = uniform(), method = "BAS", bestmodel = c(1, 
#>     1, 1, 0, 0, 1, 1, 1))
#> 
#> 
#>  Marginal Posterior Inclusion Probabilities: 
#> Intercept      npreg        glu         bp       skin        bmi        ped  
#>    1.0000     0.5337     1.0000     0.1562     0.2119     0.7425     0.8985  
#>       age  
#>    0.7236