hyper.g.n {BAS}	R Documentation

Generalized hyper-g/n Prior Distribution for g for mixtures of g-priors on Coefficients in BMA Models

Description

Creates an object representing the hyper-g/n mixture of g-priors on coefficients for BAS. This is a special case of the tCCH prior

Usage


hyper.g.n(alpha = 3, n = NULL)

Arguments

alpha
a scalar > 0, recommended 2 < alpha <= 3

n
The sample size; if NULL, the value derived from the data in the call to 'bas.glm' will be used.

Details

Creates a structure used for bas.glm. This is a special case of the tCCH, where hyper.g.n(alpha=3, n) is equivalent to tCCH(alpha=1, beta=2, s=0, r=1.5, v = 1, theta=1/n)

Value

returns an object of class "prior", with the family and hyerparameters.

Author(s)

Merlise Clyde

See Also

tCCH, robust, hyper.g, CCHbas.glm

Other beta priors: CCH, EB.local, IC.prior, Jeffreys, TG, beta.prime, g.prior, hyper.g, intrinsic, robust, tCCH, testBF.prior

Examples


n <- 500
hyper.g.n(alpha = 3, n = n)

[Package BAS version 1.5.3 Index]