intrinsic {BAS}	R Documentation

Intrinsic Prior Distribution for Coefficients in BMA Models

Description

Creates an object representing the intrinsic prior on g, a special case of the tCCH mixture of g-priors on coefficients for BAS.

Usage


intrinsic(n = NULL)

Arguments

n
the sample size; if NULL, the value derived from the data in the call to 'bas.glm' will be used.

Details

Creates a structure used for bas.glm.

Value

returns an object of class "prior", with the family "intrinsic" of class "TCCH" and hyperparameters alpha = 1, beta = 1, s = 0, r = 1, n = n for the tCCH prior where theta in the tCCH prior is determined by the model size and sample size.

Author(s)

Merlise A Clyde

References

Womack, A., Novelo,L.L., Casella, G. (2014). "Inference From Intrinsic Bayes' Procedures Under Model Selection and Uncertainty". Journal of the American Statistical Association. 109:1040-1053. http://amstat.tandfonline.com/doi/abs/10.1080/01621459.2014.880348

See Also

tCCH, robust, hyper.g, hyper.g.nbas.glm

Other beta priors: CCH, EB.local, IC.prior, Jeffreys, TG, beta.prime, g.prior, hyper.g.n, hyper.g, robust, tCCH, testBF.prior

Examples


n <- 500
tCCH(alpha = 1, beta = 2, s = 0, r = 1.5, v = 1, theta = 1 / n)

[Package BAS version 1.5.4 Index]