phi1 {BAS}	R Documentation

Compound Confluent hypergeometric function of two variables

Description

Compute the Confluent Hypergeometric function of two variables, also know as a Horn hypergeometric function or Humbert's hypergeometric used in Gordy (1998) with integral representation:

Usage


phi1(a, b, c, x, y)

Arguments

a
a > 0

b
arbitrary

c
c > 0

x
x > 0

y
0 <= y < 1

Details

phi_1(a,b,c,x,y) = Beta(a,b) Int_0^1 t^(a-1) (1 - t)^(c-a-1) (1 - yt)^(-b) exp(x t) dt https://en.wikipedia.org/wiki/Humbert_series Note that Gordy's arguments for x and y are reversed in the reference above.

Code for phi1 provided by Gordy.

Author(s)

Merlise Clyde (clyde@stat.duke.edu)

References

Gordy 1998

See Also

Other special functions: hypergeometric1F1, hypergeometric2F1

Examples



# special cases
# phi1(a, b, c, x=0, y) is the same as 2F1(b, a; c, y)
phi1(1, 2, 1.5, 0, 1 / 100)
hypergeometric2F1(2, 1, 1.5, 1 / 100, log = FALSE)

# phi1(a,0,c,x,y) is the same as 1F1(a,c,x)
phi1(1, 0, 1.5, 3, 1 / 100)
hypergeometric1F1(1, 1.5, 3, log = FALSE)

[Package BAS version 1.5.4 Index]