print.bas {BAS}	R Documentation

Print a Summary of Bayesian Model Averaging objects from BAS

Description

summary and print methods for Bayesian model averaging objects created by bas Bayesian Adaptive Sampling

Usage


## S3 method for class 'bas'
print(x, digits = max(3L, getOption("digits") - 3L), ...)

Arguments

x
object of class 'bas'

digits
optional number specifying the number of digits to display

...
other parameters to be passed to print.default

Details

The print methods display a view similar to print.lm . The summary methods display a view specific to Bayesian model averaging giving the top 5 highest probability models represented by their inclusion indicators. Summaries of the models include the Bayes Factor (BF) of each model to the model with the largest marginal likelihood, the posterior probability of the models, R2, dim (which includes the intercept) and the log of the marginal likelihood.

Author(s)

Merlise Clyde clyde@stat.duke.edu

See Also

coef.bas

Examples



library(MASS)
data(UScrime)
UScrime[, -2] <- log(UScrime[, -2])
crime.bic <- bas.lm(y ~ ., data = UScrime, n.models = 2^15, prior = "BIC", initprobs = "eplogp")
print(crime.bic)
summary(crime.bic)

[Package BAS version 1.5.3 Index]