Truncated Beta-Binomial Prior Distribution for Models

Creates an object representing the prior distribution on models for BAS using a truncated Beta-Binomial Distribution on the Model Size

Usage

tr.beta.binomial(alpha = 1, beta = 1, trunc)

Arguments

alpha

parameter in the beta prior distribution

beta

parameter in the beta prior distribution

trunc

parameter that determines truncation in the distribution i.e. P(M; alpha, beta, trunc) = 0 if M > trunc.

Value

returns an object of class "prior", with the family and hyperparameters.

Details

The beta-binomial distribution on model size is obtained by assigning each variable inclusion indicator independent Bernoulli distributions with probability w, and then giving w a beta(alpha,beta) distribution. Marginalizing over w leads to the number of included predictors having a beta-binomial distribution. The default hyperparameters lead to a uniform distribution over model size. The Truncated version assigns zero probability to all models of size > trunc.

See also

Author

Merlise Clyde

Examples

tr.beta.binomial(1, 10, 5)
#> $family
#> [1] "Trunc-Beta-Binomial"
#> 
#> $hyper.parameters
#> [1]  1 10  5
#> 
#> attr(,"class")
#> [1] "prior"
library(MASS)
data(UScrime)
UScrime[, -2] <- log(UScrime[, -2])
crime.bic <- bas.lm(y ~ .,
  data = UScrime, n.models = 2^15, prior = "BIC",
  modelprior = tr.beta.binomial(1, 1, 8),
  initprobs = "eplogp"
)