tr.poisson {BAS}	R Documentation

Truncated Poisson Prior Distribution for Models

Description

Creates an object representing the prior distribution on models for BAS using a truncated Poisson Distribution on the Model Size

Usage


tr.poisson(lambda, trunc)

Arguments

lambda
parameter in the Poisson distribution representing expected model size with infinite predictors

trunc
parameter that determines truncation in the distribution i.e. P(M; lambda, trunc) = 0 if M > trunc

Details

The Poisson prior distribution on model size is obtained by assigning each variable inclusion indicator independent Bernoulli distributions with probability w, and then taking a limit as p goes to infinity and w goes to zero, such that p*w converges to lambda. The Truncated version assigns zero probability to all models of size M > trunc.

Value

returns an object of class "prior", with the family and hyperparameters.

Author(s)

Merlise Clyde

See Also

bas.lm, Bernoulli,uniform

Other priors modelpriors: Bernoulli.heredity, Bernoulli, beta.binomial, tr.beta.binomial, tr.power.prior, uniform

Examples


tr.poisson(10, 50)

[Package BAS version 1.5.4 Index]