uniform {BAS}	R Documentation

Uniform Prior Distribution for Models

Description

Creates an object representing the prior distribution on models for BAS.

Usage


uniform()

Details

The Uniform prior distribution is a commonly used prior in BMA, and is a special case of the independent Bernoulli prior with probs=.5. The implied prior distribution on model size is binomial(p, .5).

Value

returns an object of class "prior", with the family name Uniform.

Author(s)

Merlise Clyde

See Also

bas.lm, beta.binomial,Bernoulli,

Other priors modelpriors: Bernoulli.heredity, Bernoulli, beta.binomial, tr.beta.binomial, tr.poisson, tr.power.prior

Examples


uniform()

[Package BAS version 1.5.4 Index]