which.matrix {BAS}	R Documentation

Coerce a BAS list object of models into a matrix.

Description

This function coerces the list object of models to a matrix and fill in the zeros to facilitate other computations.

Usage


which.matrix(which, n.vars)

Arguments

which
a 'bas' model object x$which

n.vars
the total number of predictors, x$n.vars

Details

which.matrix coerces x$which into a matrix.

Value

a matrix representation of x$which, with number of rows equal to the length of which.models or total number of models and number of columns x$n.vars

Author(s)

Merlise Clyde clyde@duke.edu

See Also

bas

Other as.matrix methods: list2matrix.bas, list2matrix.which

Examples



data(Hald)
Hald.bic <-  bas.lm(Y ~ ., data=Hald, prior="BIC", initprobs="eplogp")
# matrix of model indicators
models <- which.matrix(Hald.bic$which, Hald.bic$n.vars)


[Package BAS version 1.5.3 Index]