Beta-Binomial Prior Distribution for Models

Creates an object representing the prior distribution on models for BAS.

Usage

beta.binomial(alpha = 1, beta = 1)

Arguments

alpha

parameter in the beta prior distribution

beta

parameter in the beta prior distribution

Value

returns an object of class "prior", with the family and hyperparameters.

Details

The beta-binomial distribution on model size is obtained by assigning each variable inclusion indicator independent Bernoulli distributions with probability w, and then giving w a beta(alpha,beta) distribution. Marginalizing over w leads to the distribution on model size having the beta-binomial distribution. The default hyperparameters lead to a uniform distribution over model size.

See also

Author

Merlise Clyde

Examples

beta.binomial(1, 10) #' @family priors modelpriors
#> $family
#> [1] "Beta-Binomial"
#> 
#> $hyper.parameters
#> [1]  1 10
#> 
#> attr(,"class")
#> [1] "prior"