Compute Credible Intervals for BAS regression coefficients from BAS objects

Uses Monte Carlo simulations using posterior means and standard deviations of coefficients to generate draws from the posterior distributions and returns highest posterior density (HPD) credible intervals. If the number of models equals one, then use the t distribution to find intervals. These currently condition on the estimate of $g$. than the description above ~~

# S3 method for coef.bas
confint(object, parm, level = 0.95, nsim = 10000, ...)

Arguments

object
a coef.bas object

parm
a specification of which parameters are to be given credible intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level
the probability coverage required

nsim
number of Monte Carlo draws from the posterior distribution. Used when number of models is greater than 1.

...
other arguments to passed; none currently

Value

A matrix (or vector) with columns giving lower and upper HPD credible limits for each parameter. These will be labeled as 1-level)/2 and 1 - (1-level)/2 in percent (by default 2.5 and 97.5).

Note

For mixture of g-priors these are approximate. This uses Monte Carlo sampling so results may be subject to Monte Carlo variation and larger values of nsim may be needed to reduce variability.

See also

Examples

data("Hald")
hald_gprior <-  bas.lm(Y~ ., data=Hald, alpha=13,
                            prior="g-prior")
coef_hald <- coef(hald_gprior)
confint(coef_hald)

#>                2.5%      97.5%       beta
#> Intercept 93.880781 96.9113634 95.4230769
#> X1         0.000000  1.8942471  1.2150202
#> X2        -1.094706  0.9383005  0.2756235
#> X3        -1.490076  0.6258418 -0.1270575
#> X4        -1.721432  0.2611582 -0.3268710
#> attr(,"Probability")
#> [1] 0.95
#> attr(,"class")
#> [1] "confint.bas"

confint(coef_hald, approx=FALSE, nsim=5000)

#>                2.5%      97.5%       beta
#> Intercept 93.910853 97.0365545 95.4230769
#> X1         0.000000  1.8645339  1.2150202
#> X2        -1.111524  0.8959157  0.2756235
#> X3        -1.562010  0.5333807 -0.1270575
#> X4        -1.675039  0.2779912 -0.3268710
#> attr(,"Probability")
#> [1] 0.95
#> attr(,"class")
#> [1] "confint.bas"

# extract just the coefficient of X4
confint(coef_hald, parm="X4")

#>         2.5%     97.5%      beta
#> X4 -1.661371 0.3088764 -0.326871
#> attr(,"Probability")
#> [1] 0.95
#> attr(,"class")
#> [1] "confint.bas"