hypergeometric1F1 {BAS}	R Documentation

Confluent hypergeometric2F1 function

Description

Compute the Confluent Hypergeometric function: 1F1(a,b,c,t) = Gamma(b)/(Gamma(b-a)Gamma(a)) Int_0^1 t^(b-1) (1 - t)^(b-a-1) exp(c t) dt

Usage


hypergeometric1F1(a, b, c, laplace = FALSE, log = TRUE)

Arguments

a
arbitrary

b
Must be greater 0

c
arbitrary

laplace
The default is to use the Cephes library; for large a or s this may return an NA, Inf or negative values,, in which case you should use the Laplace approximation.

log
if TRUE, return log(1F1)

Author(s)

Merlise Clyde (clyde@stat.duke.edu)

References

Cephes library hyp1f1.c

See Also

Other special functions: hypergeometric2F1, phi1

Examples


hypergeometric1F1(11.14756, 0.5, 0.00175097)



[Package BAS version 1.5.3 Index]