Intrinsic Prior Distribution for Coefficients in BMA Models

Creates an object representing the intrinsic prior on g, a special case of the tCCH mixture of g-priors on coefficients for BAS.

Usage

intrinsic(n = NULL)

Arguments

n

the sample size; if NULL, the value derived from the data in the call to `bas.glm` will be used.

Value

returns an object of class "prior", with the family "intrinsic" of class "TCCH" and hyperparameters alpha = 1, beta = 1, s = 0, r = 1, n = n for the tCCH prior where theta in the tCCH prior is determined by the model size and sample size.

Details

Creates a structure used for bas.glm.

References

Womack, A., Novelo,L.L., Casella, G. (2014). "Inference From Intrinsic Bayes' Procedures Under Model Selection and Uncertainty". Journal of the American Statistical Association. 109:1040-1053. doi:10.1080/01621459.2014.880348

See also

Author

Merlise A Clyde

Examples

n <- 500
tCCH(alpha = 1, beta = 2, s = 0, r = 1.5, v = 1, theta = 1 / n)
#> $family
#> [1] "tCCH"
#> 
#> $class
#> [1] "TCCH"
#> 
#> $hyper.parameters
#> $hyper.parameters$alpha
#> [1] 1
#> 
#> $hyper.parameters$beta
#> [1] 2
#> 
#> $hyper.parameters$s
#> [1] 0
#> 
#> $hyper.parameters$r
#> [1] 1.5
#> 
#> $hyper.parameters$v
#> [1] 1
#> 
#> $hyper.parameters$theta
#> [1] 0.002
#> 
#> 
#> attr(,"class")
#> [1] "prior"