Compound Confluent hypergeometric function of two variables

Compute the Confluent Hypergeometric function of two variables, also know as a Horn hypergeometric function or Humbert's hypergeometric used in Gordy (1998) with integral representation:

phi1(a, b, c, x, y)

Arguments

a
a > 0

b
arbitrary

c
c > 0

x
x > 0

y
0 <= y < 1

Details

phi_1(a,b,c,x,y) = Beta(a,b) Int_0^1 t^(a-1) (1 - t)^(c-a-1) (1 - yt)^(-b) exp(x t) dt https://en.wikipedia.org/wiki/Humbert_series Note that Gordy's arguments for x and y are reversed in the reference above.

Code for phi1 provided by Gordy.

References

Gordy 1998

See also

Examples

# special cases
# Phi1(a, b, c, x=0, y) = 2F1(b, a; c, y)
phi1(1, 2, 1.5, 0, 1/100);

#> [1] 1.013495

hypergeometric2F1(2, 1, 1.5, 1/100, log = FALSE)

#> [1] 1.013495

# Phi1(a, b=0, c, x, y) = Phi(a, b, c, x, y=0) = 1F1(a, c, x) ## ??
phi1(1, 0, 1.5, 3, 1/100);

#> [1] 10.13001

hypergeometric1F1(1, 1.5, 3, log = FALSE);

#> [1] 10.13001